avoid some compiler warnings
git-svn-id: https://svn.r-project.org/R/trunk@82803 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/src/gramRd.c
@@ -3237,7 +3237,7 @@ static SEXP xxnewcommand(SEXP cmd, SEXP name, SEXP defn, YYLTYPE *lloc)
     {
         if (TYPEOF(val = CAR(defnvals)) == STRSXP)
         {
-            vlen = strlen(CHAR(STRING_ELT(val, 0)));
+            vlen = (int)strlen(CHAR(STRING_ELT(val, 0)));
             defnBuffer = R_Realloc(defnBuffer, len + vlen + 1, char);
             strncpy(defnBuffer + len, CHAR(STRING_ELT(val, 0)), vlen + 1);
             len += vlen;
